Fix backup and restore scripts
dbuser is a Unix user in the Docker image, so it looks right. $DB_USER
contains the username ("role") that's configured in the database itself.

Removed the ability to automatically restore from an unspecified backup
file. That's dangerous. If we want this logic to find the latest backup
file, we should be doing it in a calling script, not as a default
behavior!